requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
schedule>=1.0.0
openai
python-dotenv
selenium
webdriver-manager